ai/requirements.txt
# Core ML
torch==2.2.0
torchvision==0.17.0
torch-geometric==2.5.0
torch-scatter
torch-sparse
torch-cluster

# ONNX
onnxruntime==1.17.0
onnx==1.15.0

# Geospatial
rasterio==1.3.9
pyproj
shapely

# Data
numpy>=1.24
pandas>=2.0
scipy>=1.11

# API / async
requests
aiohttp
python-dotenv

# Training UI
fastapi>=0.110
uvicorn[standard]
jinja2
aiofiles

# Utilities
tqdm
matplotlib
scikit-learn
